start moving field to shared component
